requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
colorama>=0.4.6
